+++ main/core/Resources/modules/resource/actions/export.js
@@ -1,5 +1,6 @@
 import {trans} from '#/main/core/translation'
 import {ASYNC_BUTTON} from '#/main/app/buttons'
+import {url} from '#/main/app/api'
 
 const action = (resourceNodes) => ({ // todo collection
   name: 'export',
@@ -7,11 +8,10 @@ const action = (resourceNodes) => ({ // todo collection
   icon: 'fa fa-fw fa-download',
   label: trans('export', {}, 'actions'),
   request: {
-    url: ['claro_resource_action', {
-      type: resourceNodes[0].meta.type,
-      action: 'export',
-      id: resourceNodes[0].id
-    }]
+    url: url(
+      ['claro_resource_download'],
+      {ids: resourceNodes.map(resourceNode => resourceNode.id)}
+    )
   }
 })
